[Change] Fix UUID useability

+++ Meduza/Include/Core/GUID.h
@@ -6,11 +6,11 @@ namespace Me
 	{
 	public:
 		UUID();
-		UUID(uint64_t const a_id);
+		UUID(uint32_t const a_id);
 		UUID(UUID const&) = default;
 
-		operator uint64_t() const { return m_guid; }
+		operator uint32_t() const { return m_guid; }
 	private:
-		uint64_t m_guid;
+		uint32_t m_guid;
 	};
 }
