+++ Meduza/Include/ECS/EntityTypes.h
@@ -1,8 +1,8 @@
 #pragma once
 
-using EntityID = uint64_t;
-using ComponentID = uint64_t;
-using SystemID = uint64_t;
+using EntityID = uint32_t;
+using ComponentID = uint32_t;
+using SystemID = uint32_t;
 
 using EntityFilter = std::set<ComponentID>;
