+++ Meduza/Include/Platform/General/Editor/EditorHelper.h
@@ -2,7 +2,7 @@
 
 namespace Me
 {
-    using EntityID = uint64_t;
+    using EntityID = uint32_t;
 
     namespace Editor
     {
