+++ Meduza/Include/Platform/General/Editor/EditorViewport.h
@@ -4,7 +4,7 @@
 
 namespace Me
 {
-    using EntityID = uint64_t;
+    using EntityID = uint32_t;
 
     namespace Renderer
     {
